requirements.txt
Django==6.0.2
djangorestframework==3.15.2
Pillow==12.1.1
